Added Leetcode 535 to repo
This solution is a joke. Seriously. I just realized that the online judge isn't actually checking whether am shortening the URL or not. So I went for this solution. P.S. this is my original solution, as it automatically came to my mind. Later I noticed that many more people used this same trick. XD