Added Leetcode 284 to repo.
Converted the Iterator to complete list. Then made a ListIterator on this list, and used the previous() method to implement the peek() method. Could've just copied the iterator to a new variable, and then implemented the peek() method using this new iterator, without touching the original one. But nvm, original submission. :)